refactor(cockpit): extract review map proof rendering
