+++ crates/tokmd-cockpit/src/render.rs
@@ -12,6 +12,7 @@ mod evidence;
 mod manifest;
 mod markdown;
 mod review_map;
+mod review_map_proof;
 mod review_packet;
 mod sections;
